+++ composer.json
@@ -1,6 +1,6 @@
 {
     "name": "roundcube/plugin-installer",
-    "description": "A composer-installer for Roundcube plugins.",
+    "description": "A composer-installer for Roundcube plugins and skins.",
     "type": "composer-installer",
     "license": "GPL-3.0+",
     "authors": [
@@ -11,6 +11,10 @@
         {
             "name": "Till Klampaeckel",
             "email": "till@php.net"
+        },
+        {
+            "name": "Philip Weir",
+            "email": "roundcube@tehinterweb.co.uk"
         }
     ],
     "autoload": {
@@ -19,7 +23,10 @@
         }
     },
     "extra": {
-        "class": "Roundcube\\Composer\\PluginInstaller"
+        "class": [
+            "Roundcube\\Composer\\PluginInstaller",
+            "Roundcube\\Composer\\SkinInstaller"
+        ]
     },
     "require": {
         "php": ">=5.3.0"
